Seperated the pages into files Completed bottom nav component and its navigation
