--- app/src/main/java/com/example/assignment_1/DietChartPage.kt
@@ -1,4 +1,0 @@
-package com.example.assignment_1
-
-class DietChartPage {
-}
