+++ app/src/main/java/com/example/assignment_1/MainActivity.kt
@@ -135,8 +135,9 @@ const val LOGIN_SCREEN = "login"
 const val SIGNUP_SCREEN = "signup"
 const val HOME_SCREEN = "home"
 const val SETTINGS_SCREEN = "settings"
-const val MEAL_PLANNER_SCREEN = "planner"
+const val MEAL_RECIPE_SCREEN = "recipe"
 const val USER_PROFILE_SCREEN = "userProfile"
+const val DIET_PLANNER = "planner"
 const val FORGOT_PASSWORD_SCREEN = "FORGOT_PASSWORD"
 
 //Database Constants
@@ -174,6 +175,7 @@ fun getPref(key: String, type: Class<Any>, context: Context) {
 }
 
 
+@RequiresApi(Build.VERSION_CODES.O)
 @Composable
 fun AppContent() {
     val navController = rememberNavController()
@@ -195,6 +197,15 @@ fun AppContent() {
         }
         composable(FORGOT_PASSWORD_SCREEN) {
             ForgotPasswordScreen(navController = navController)
+        }
+        composable(SETTINGS_SCREEN){
+            SettingsScreen(navController = navController, viewModel = SettingsViewModel())
+        }
+        composable(MEAL_RECIPE_SCREEN){
+            DietSelectionPage(navController = navController)
+        }
+        composable(DIET_PLANNER){
+            DisplayDatePicker(navController = navController)
         }
 
     }
@@ -724,123 +735,6 @@ fun FeatureRow(icon: ImageVector, text: String) {
     }
 }
 
-@Composable
-fun HomeScreen(navController: NavController) {
-    Surface(color = MaterialTheme.colorScheme.surface) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(16.dp),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Text(
-                text = "MealMate - Plan, Cook, Eat.",
-                textAlign = TextAlign.Center,
-                style = TextStyle(
-                    fontWeight = FontWeight.Bold,
-                    fontSize = 32.sp, // Adjust size as needed
-                    color = MaterialTheme.colorScheme.onSurface
-                )
-            )
-            Text(
-                text = "Streamline your meals, save time & money, and eat healthy!",
-                textAlign = TextAlign.Center,
-                modifier = Modifier.padding(vertical = 8.dp)
-            )
-
-            // Plan your week section
-            Text(
-                text = "Plan your week in minutes:",
-                textAlign = TextAlign.Center,
-                modifier = Modifier.padding(vertical = 16.dp)
-            )
-
-            // Carousel of cuisines
-            // Video clip of drag-and-drop
-            // Image of grocery list
-            // Features section
-            Text(
-                text = "Features you'll love:",
-                textAlign = TextAlign.Center,
-                modifier = Modifier.padding(vertical = 16.dp)
-            )
-
-            // Feature icons with text
-            FeatureRow(
-                icon = Icons.Default.Timer,
-                text = "30-minute meals: Quick and easy recipes to fit your busy schedule."
-            )
-            FeatureRow(
-                icon = Icons.Default.Filter,
-                text = "Dietary filters: Find meals that meet your specific needs."
-            )
-            FeatureRow(
-                icon = Icons.Default.RestaurantMenu,
-                text = "Leftover magic: Repurpose ingredients to avoid food waste."
-            )
-            FeatureRow(
-                icon = Icons.Default.FamilyRestroom,
-                text = "Family-friendly options: Keep everyone happy with kid-approved meals."
-            )
-            FeatureRow(
-                icon = Icons.Default.Phonelink,
-                text = "Offline access: Take your meal plan anywhere."
-            )
-
-            // Get started section
-            Text(
-                text = "Get started today!",
-                textAlign = TextAlign.Center,
-                modifier = Modifier.padding(vertical = 16.dp)
-            )
-            MyButton("View Diet Options", {})
-            MyButton("View Meal Plans", {})
-            MyButton("View Nutrition Chart", {})
-
-            Box(modifier = Modifier.fillMaxSize()) {
-                // add your column here (with align modifier)
-                Column(modifier = Modifier.align(Alignment.BottomCenter)) {
-                    BottomNavigation(
-                        modifier = Modifier.fillMaxWidth(),
-                        backgroundColor = MaterialTheme.colorScheme.surface
-                    ) {
-                        BottomNavigationItem(
-                            selected = navController.currentDestination?.route == "mealList",
-                            onClick = { navController.navigate("mealList") },
-                            icon = { Icon(Icons.Default.Search, contentDescription = "Meals") },
-                            label = { Text("Meals") }
-                        )
-                        BottomNavigationItem(
-                            selected = navController.currentDestination?.route == "shoppingList",
-                            onClick = { navController.navigate("shoppingList") },
-                            icon = {
-                                Icon(
-                                    Icons.Default.LocalPizza,
-                                    contentDescription = "Diet Planner"
-                                )
-                            },
-                            label = { Text("Diet Planner") }
-                        )
-                        BottomNavigationItem(
-                            selected = navController.currentDestination?.route == "settings",
-                            onClick = { navController.navigate("settings") },
-                            icon = {
-                                Icon(
-                                    Icons.Default.Settings,
-                                    contentDescription = "Settings"
-                                )
-                            },
-                            label = { Text("Settings") }
-                        )
-                    }
-
-                }
-            }
-        }
-    }
-}
-
 @Preview(showBackground = true)
 @Composable
 fun HomeScreenPreview() {
@@ -848,105 +742,6 @@ fun HomeScreenPreview() {
     HomeScreen(navController)
 }
 
-@Composable
-fun SettingsScreen(navController: NavController, viewModel: SettingsViewModel) {
-    Surface(color = MaterialTheme.colorScheme.surface) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(16.dp)
-        ) {
-            Text(text = "Settings", style = MaterialTheme.typography.titleLarge)
-            Spacer(modifier = Modifier.height(16.dp))
-            LazyColumn {
-                items(viewModel.settingsItems) { setting ->
-                    Box(
-                        modifier = Modifier
-                            .fillMaxWidth() // Fill the width of the column
-                            .height(72.dp) // Adjust the height as needed
-                    ) {
-                        SettingsItem(setting = setting)
-                    }
-                    Divider()
-                }
-            }
-            Spacer(modifier = Modifier.weight(1f))
-            Button(
-                onClick = { /* Handle save button click */ },
-                modifier = Modifier.align(Alignment.End)
-            ) {
-                Text(text = "Save")
-            }
-        }
-        Box(modifier = Modifier.fillMaxSize()) {
-            // add your column here (with align modifier)
-            Column(modifier = Modifier.align(Alignment.BottomCenter)) {
-                BottomNavigation(
-                    modifier = Modifier.fillMaxWidth(),
-                    backgroundColor = MaterialTheme.colorScheme.surface
-                ) {
-                    BottomNavigationItem(
-                        selected = navController.currentDestination?.route == "mealList",
-                        onClick = { navController.navigate("mealList") },
-                        icon = { Icon(Icons.Default.Search, contentDescription = "Meals") },
-                        label = { Text("Meals") }
-                    )
-                    BottomNavigationItem(
-                        selected = navController.currentDestination?.route == "shoppingList",
-                        onClick = { navController.navigate("shoppingList") },
-                        icon = {
-                            Icon(
-                                Icons.Default.LocalPizza,
-                                contentDescription = "Diet Planner"
-                            )
-                        },
-                        label = { Text("Diet Planner") }
-                    )
-                    BottomNavigationItem(
-                        selected = navController.currentDestination?.route == "settings",
-                        onClick = { navController.navigate("settings") },
-                        icon = { Icon(Icons.Default.Settings, contentDescription = "Settings") },
-                        label = { Text("Settings") }
-                    )
-                }
-
-            }
-        }
-    }
-}
-
-@Composable
-fun SettingsItem(setting: String) {
-    Text(
-        text = setting,
-        style = MaterialTheme.typography.bodyMedium.copy(fontSize = 16.sp),
-        modifier = Modifier
-            .padding(vertical = 16.dp, horizontal = 16.dp)
-            .fillMaxWidth()
-    )
-}
-
-class SettingsViewModel : ViewModel() {
-    val settingsItems = listOf(
-        "Notification",
-        "Sound",
-        "Dark Mode",
-        "Language",
-        "Theme Color",
-        "Font Size",
-        "Privacy",
-        "Security",
-        "Data Usage",
-        "Account",
-        // Add more settings items as needed
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun SettingsScreenPreview() {
-    SettingsScreen(navController = rememberNavController(), viewModel = SettingsViewModel())
-}
 
 @Composable
 fun MyButton(
@@ -974,171 +769,6 @@ fun MyButton(
     }
 }
 
-@Composable
-fun NutritionPage(
-    navController: NavController,
-    foodName: String,
-    ingredients: List<String>,
-    nutritionValues: Map<String, Float>
-) {
-    val paint = remember {
-        Paint().apply {
-            color = Color.Black.toArgb() // Set paint color
-            textAlign = Paint.Align.CENTER // Set text alignment
-            textSize = 24f
-        }
-    }
-    val nutrientColors = remember {
-        val colorPalette = listOf(
-            Color.Blue,
-            Color.Green,
-            Color.Red,
-            Color.Yellow,
-            Color.Magenta,
-            Color.Cyan
-        )
-        nutritionValues.keys.zip(colorPalette).toMap()
-    }
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp)
-    ) {
-        Text(
-            text = "Nutrition Distribution",
-            style = MaterialTheme.typography.headlineLarge,
-            modifier = Modifier.padding(bottom = 16.dp)
-        )
-        val navController = rememberNavController()
-        NutritionPieChart(nutritionValues, ingredients, paint, nutrientColors)
-        NutritionLegend(nutritionValues, nutrientColors)
-        Spacer(modifier = Modifier.height(16.dp))
-        Text(
-            text = "Food Name: $foodName",
-            style = MaterialTheme.typography.headlineMedium
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-        Text(
-            text = "Ingredients: ${ingredients.joinToString(", ")}",
-            style = MaterialTheme.typography.bodyMedium
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-        Text(
-            text = "Nutrition Values",
-            style = MaterialTheme.typography.headlineMedium
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-        NutritionTable(nutritionValues)
-        Box(modifier = Modifier.fillMaxSize()) {
-            // add your column here (with align modifier)
-            Column(modifier = Modifier.align(Alignment.BottomCenter)) {
-                BottomNavigation(
-                    modifier = Modifier.fillMaxWidth(),
-                    backgroundColor = MaterialTheme.colorScheme.surface
-                ) {
-                    BottomNavigationItem(
-                        selected = navController.currentDestination?.route == "mealList",
-                        onClick = { navController.navigate("mealList") },
-                        icon = { Icon(Icons.Default.Search, contentDescription = "Meals") },
-                        label = { Text("Meals") }
-                    )
-                    BottomNavigationItem(
-                        selected = navController.currentDestination?.route == "shoppingList",
-                        onClick = { navController.navigate("shoppingList") },
-                        icon = {
-                            Icon(
-                                Icons.Default.LocalPizza,
-                                contentDescription = "Diet Planner"
-                            )
-                        },
-                        label = { Text("Diet Planner") }
-                    )
-                    BottomNavigationItem(
-                        selected = navController.currentDestination?.route == "settings",
-                        onClick = { navController.navigate("settings") },
-                        icon = { Icon(Icons.Default.Settings, contentDescription = "Settings") },
-                        label = { Text("Settings") }
-                    )
-                }
-
-            }
-        }
-    }
-}
-
-
-@Composable
-fun NutritionPieChart(
-    nutritionValues: Map<String, Float>,
-    ingredients: List<String>,
-    paint: Paint,
-    nutrientColors: Map<String, Color>
-) {
-    val totalValue = nutritionValues.values.sum()
-    Canvas(
-        modifier = Modifier
-            .fillMaxWidth()
-            .aspectRatio(1f)
-            .border(width = 1.dp, color = Color.Black)
-    ) {
-        var startAngle = 0f
-        nutritionValues.forEach { (label, value) ->
-            val sweepAngle = (value / totalValue) * 360
-            val ingredientIndex = ingredients.indexOf(label)
-            val ingredientName = if (ingredientIndex != -1) ingredients[ingredientIndex] else ""
-            drawArc(
-                color = nutrientColors[label] ?: Color.Gray,
-                startAngle = startAngle,
-                sweepAngle = sweepAngle,
-                useCenter = true,
-                topLeft = Offset(size.width / 4, size.height / 4),
-                size = Size(size.width / 2, size.height / 2),
-                style = Stroke(width = 40f)
-            )
-            if (ingredientName.isNotBlank()) {
-                val angle = Math.toRadians(startAngle.toDouble() + sweepAngle.toDouble() / 2)
-                val textX = center.x + size.width / 4 * cos(angle).toFloat() - 20.dp.toPx()
-                val textY = center.y + size.height / 4 * sin(angle).toFloat()
-                drawIntoCanvas {
-                    it.nativeCanvas.drawText(
-                        ingredientName,
-                        textX,
-                        textY,
-                        paint
-                    )
-                }
-            }
-            startAngle += sweepAngle
-        }
-    }
-}
-
-
-@Composable
-fun NutritionTable(nutritionValues: Map<String, Float>) {
-    Column {
-        nutritionValues.forEach { (nutrient, value) ->
-            Row {
-                Text(text = nutrient, modifier = Modifier.weight(1f))
-                Text(text = ": $value", modifier = Modifier.weight(1f))
-            }
-        }
-    }
-}
-
-@Preview
-@Composable
-fun PreviewNutritionPage() {
-    val nutritionValues = mapOf(
-        "Protein" to 20f,
-        "Carbohydrates" to 50f,
-        "Fat" to 30f
-    )
-    val foodName = "Sample Food"
-    val ingredients = listOf("Ingredient 1", "Ingredient 2", "Ingredient 3")
-    val navController = rememberNavController()
-    NutritionPage(navController, foodName, ingredients, nutritionValues)
-}
 
 @Composable
 fun NutritionLegend(
@@ -1170,417 +800,6 @@ fun RecipeItem(recipeName: String) {
             .fillMaxWidth()
             .padding(vertical = 8.dp)
     )
-}
-
-@RequiresApi(Build.VERSION_CODES.O)
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun DietSelectionPage(onDietSelected: (String) -> Unit, navController: NavController) {
-    val veganRecipes = listOf(
-        "LENTIL BOLOGNESE",
-        "BUTTERNUT SQUASH RISOTTO",
-        "BOMBAY BURRITOS"
-    )
-    val ketoRecipes = listOf(
-        "BACON-WRAPPED AVOCADO",
-        "KETO CHICKEN ALFREDO",
-        "SPINACH AND FETA STUFFED SALMON"
-    )
-    val vegetarianRecipes = listOf(
-        "VEGETARIAN PAD THAI",
-        "DOSA WITH SAMBAR",
-        "MUSHROOM RISOTTO"
-    )
-    val highProteinRecipes = listOf(
-        "GRILLED CHICKEN BREASTS",
-        "TURKEY AND QUINOA STUFFED BELL PEPPERS",
-        "SALMON AND ASPARAGUS FOIL PACKS"
-    )
-    var selectedRecepie = ""
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp)
-    ) {
-        // Diet selection dropdown
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(color = Color.LightGray, shape = RoundedCornerShape(8.dp))
-                .padding(vertical = 8.dp, horizontal = 16.dp)
-        ) {
-            Text(
-                text = "Select Diet for Recepies",
-                style = MaterialTheme.typography.titleLarge,
-                modifier = Modifier.align(Alignment.CenterStart)
-            )
-        }
-
-        val mealPlans = listOf(
-            "Vegan", "Keto", "Vegetarian", "High Protein"
-        )
-        var isExpanded by remember { mutableStateOf(false) }
-        var selectedMealPlanState by remember { mutableStateOf(mealPlans[0]) }
-
-        Column(modifier = Modifier.padding(16.dp)) {
-            Text(
-                text = "Choose your Meal Plan",
-                style = MaterialTheme.typography.labelLarge
-            )
-
-            ExposedDropdownMenuBox(
-                expanded = isExpanded,
-                onExpandedChange = { isExpanded = it },
-            ) {
-                TextField(
-                    modifier = Modifier
-                        .menuAnchor()
-                        .fillMaxWidth()
-                        .focusProperties { canFocus = false }
-                        .padding(bottom = 8.dp),
-                    readOnly = true,
-                    value = selectedMealPlanState,
-                    onValueChange = {},
-                    label = { Text("Meal Plan") },
-                    trailingIcon = {
-                        ExposedDropdownMenuDefaults.TrailingIcon(expanded = isExpanded)
-                    }
-                )
-                ExposedDropdownMenu(
-                    expanded = isExpanded,
-                    onDismissRequest = { isExpanded = false }
-                ) {
-                    mealPlans.forEach { selectionOption ->
-                        DropdownMenuItem(
-                            text = { Text(selectionOption) },
-                            onClick = {
-                                selectedMealPlanState = selectionOption
-                                isExpanded = false
-                            },
-                            contentPadding = ExposedDropdownMenuDefaults.ItemContentPadding,
-                        )
-                    }
-                }
-            }
-            selectedRecepie = selectedMealPlanState
-//            Text(selectedRecepie)
-            Spacer(modifier = Modifier.height(16.dp))
-            when (selectedRecepie) {
-                "Vegan" -> {
-                    LazyColumn {
-                        items(veganRecipes) { recipe ->
-                            RecipeItem(recipeName = recipe)
-                        }
-                    }
-                }
-
-                "Keto" -> {
-                    LazyColumn {
-                        items(ketoRecipes) { recipe ->
-                            RecipeItem(recipeName = recipe)
-                        }
-                    }
-                }
-
-                "Vegetarian" -> {
-                    LazyColumn {
-                        items(vegetarianRecipes) { recipe ->
-                            RecipeItem(recipeName = recipe)
-                        }
-                    }
-                }
-
-                "High Protein" -> {
-                    LazyColumn {
-                        items(highProteinRecipes) { recipe ->
-                            RecipeItem(recipeName = recipe)
-                        }
-                    }
-                }
-
-            }
-        }
-    }
-    Box(modifier = Modifier.fillMaxSize()) {
-        // add your column here (with align modifier)
-        Column(modifier = Modifier.align(Alignment.BottomCenter)) {
-            BottomNavigation(
-                modifier = Modifier.fillMaxWidth(),
-                backgroundColor = MaterialTheme.colorScheme.surface
-            ) {
-                BottomNavigationItem(
-                    selected = navController.currentDestination?.route == "mealList",
-                    onClick = { navController.navigate("mealList") },
-                    icon = { Icon(Icons.Default.Search, contentDescription = "Meals") },
-                    label = { Text("Meals") }
-                )
-                BottomNavigationItem(
-                    selected = navController.currentDestination?.route == "shoppingList",
-                    onClick = { navController.navigate("shoppingList") },
-                    icon = { Icon(Icons.Default.LocalPizza, contentDescription = "Diet Planner") },
-                    label = { Text("Diet Planner") }
-                )
-                BottomNavigationItem(
-                    selected = navController.currentDestination?.route == "settings",
-                    onClick = { navController.navigate("settings") },
-                    icon = { Icon(Icons.Default.Settings, contentDescription = "Settings") },
-                    label = { Text("Settings") }
-                )
-            }
-
-        }
-    }
-}
-
-
-@RequiresApi(Build.VERSION_CODES.O)
-@Preview
-@Composable
-fun PreviewDietSelectionPage() {
-    val navController = rememberNavController()
-    DietSelectionPage({}, navController)
-}
-
-@RequiresApi(Build.VERSION_CODES.O)
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun DisplayDatePicker(navController: NavController) {
-    val calendar = Calendar.getInstance()
-    calendar.set(2024, 0, 1) // month (0) is January
-    val datePickerState = rememberDatePickerState(
-        initialSelectedDateMillis = Instant.now().toEpochMilli()
-    )
-    var showDatePicker by remember { mutableStateOf(false) }
-    var selectedDate by rememberSaveable { mutableLongStateOf(calendar.timeInMillis) }
-
-    Column(modifier = Modifier.padding(16.dp)) {
-        Text(text = "Meal Planner", fontSize = 30.sp, style = MaterialTheme.typography.bodyMedium)
-        Spacer(modifier = Modifier.height(16.dp))
-        Text(
-            text = "Set your target, Follow your Diet",
-            style = MaterialTheme.typography.headlineSmall
-        )
-        Spacer(modifier = Modifier.height(16.dp))
-        if (showDatePicker) {
-            DatePickerDialog(
-                onDismissRequest = { showDatePicker = false },
-                confirmButton = {
-                    TextButton(onClick = {
-                        showDatePicker = false
-                        selectedDate = datePickerState.selectedDateMillis!!
-                    }) {
-                        Text(
-                            text = "OK",
-                            modifier = Modifier.padding(16.dp), // Add the modifier parameter
-                            style = MaterialTheme.typography.headlineSmall
-                        )
-                    }
-                },
-                dismissButton = {
-                    TextButton(onClick = { showDatePicker = false }) {
-                        Text(
-                            text = "Cancel",
-                            modifier = Modifier.padding(16.dp), // Add the modifier parameter
-                            style = MaterialTheme.typography.headlineSmall
-                        )
-                    }
-                }
-            ) {
-                DatePicker(state = datePickerState)
-            }
-        }
-        Button(
-            onClick = { showDatePicker = true }
-        ) {
-            Text(
-                text = "Enter Date of target",
-                modifier = Modifier.padding(16.dp), // Add the modifier parameter
-                style = MaterialTheme.typography.headlineMedium
-            )
-        }
-        val formatter = SimpleDateFormat("dd/MM/yyyy", Locale.ROOT)
-        Text(
-            text = "Date of target: ${formatter.format(Date(selectedDate))}"
-        )
-
-        val mealPlans = listOf(
-            "Celebration or Seasonal Meal Plans",
-            "Standard Meal Plans",
-            "Specialized Meal Plans",
-            "Cuisine-Based Meal Plans",
-            "Meal Plans for Specific Goals",
-            "Customizable Meal Plans",
-            "Celebration or Seasonal Meal Plans"
-        )
-        var isExpanded by remember { mutableStateOf(false) }
-        var selectedMealPlanState by remember { mutableStateOf(mealPlans[0]) }
-
-        Column(modifier = Modifier.padding(16.dp)) {
-            Text(
-                text = "Choose your Meal Plan",
-                style = MaterialTheme.typography.labelLarge
-            )
-
-            ExposedDropdownMenuBox(
-                expanded = isExpanded,
-                onExpandedChange = { isExpanded = it },
-            ) {
-                TextField(
-                    modifier = Modifier
-                        .menuAnchor()
-                        .fillMaxWidth()
-                        .focusProperties { canFocus = false }
-                        .padding(bottom = 8.dp),
-                    readOnly = true,
-                    value = selectedMealPlanState,
-                    onValueChange = {},
-                    label = { Text("Meal Plan") },
-                    trailingIcon = {
-                        ExposedDropdownMenuDefaults.TrailingIcon(expanded = isExpanded)
-                    }
-                )
-                ExposedDropdownMenu(
-                    expanded = isExpanded,
-                    onDismissRequest = { isExpanded = false }
-                ) {
-                    mealPlans.forEach { selectionOption ->
-                        DropdownMenuItem(
-                            text = { Text(selectionOption) },
-                            onClick = {
-                                selectedMealPlanState = selectionOption
-                                isExpanded = false
-                            },
-                            contentPadding = ExposedDropdownMenuDefaults.ItemContentPadding,
-                        )
-                    }
-                }
-            }
-        }
-
-    }
-    Box(modifier = Modifier.fillMaxSize()) {
-        // add your column here (with align modifier)
-        Column(modifier = Modifier.align(Alignment.BottomCenter)) {
-            BottomNavigation(
-                modifier = Modifier.fillMaxWidth(),
-                backgroundColor = MaterialTheme.colorScheme.surface
-            ) {
-                BottomNavigationItem(
-                    selected = navController.currentDestination?.route == "mealList",
-                    onClick = { navController.navigate("mealList") },
-                    icon = { Icon(Icons.Default.Search, contentDescription = "Meals") },
-                    label = { Text("Meals") }
-                )
-                BottomNavigationItem(
-                    selected = navController.currentDestination?.route == "shoppingList",
-                    onClick = { navController.navigate("shoppingList") },
-                    icon = { Icon(Icons.Default.LocalPizza, contentDescription = "Diet Planner") },
-                    label = { Text("Diet Planner") }
-                )
-                BottomNavigationItem(
-                    selected = navController.currentDestination?.route == "settings",
-                    onClick = { navController.navigate("settings") },
-                    icon = { Icon(Icons.Default.Settings, contentDescription = "Settings") },
-                    label = { Text("Settings") }
-                )
-            }
-
-        }
-    }
-}
-
-@RequiresApi(Build.VERSION_CODES.O)
-@Preview(showBackground = true)
-@Composable
-private fun PreviewDisplayDatePicker() {
-    val navController = rememberNavController()
-    DisplayDatePicker(navController)
-}
-
-@Composable
-fun ProfileScreen(navController: NavController, viewModel: ProfileViewModel) {
-    Surface(color = MaterialTheme.colorScheme.surface) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(16.dp)
-        ) {
-            Text(text = "Profile", style = MaterialTheme.typography.bodyMedium)
-            Spacer(modifier = Modifier.height(16.dp))
-            LazyColumn {
-                items(viewModel.profileItem) { setting ->
-                    ProfileItem(setting = setting)
-                    Divider()
-                }
-            }
-            Spacer(modifier = Modifier.weight(1f))
-            Button(
-                onClick = { /* Handle save button click */ },
-                modifier = Modifier.align(Alignment.End)
-            ) {
-                Text(text = "Save")
-            }
-        }
-    }
-    Box(modifier = Modifier.fillMaxSize()) {
-        // add your column here (with align modifier)
-        Column(modifier = Modifier.align(Alignment.BottomCenter)) {
-            BottomNavigation(
-                modifier = Modifier.fillMaxWidth(),
-                backgroundColor = MaterialTheme.colorScheme.surface
-            ) {
-                BottomNavigationItem(
-                    selected = navController.currentDestination?.route == "mealList",
-                    onClick = { navController.navigate("mealList") },
-                    icon = { Icon(Icons.Default.Search, contentDescription = "Meals") },
-                    label = { Text("Meals") }
-                )
-                BottomNavigationItem(
-                    selected = navController.currentDestination?.route == "shoppingList",
-                    onClick = { navController.navigate("shoppingList") },
-                    icon = { Icon(Icons.Default.LocalPizza, contentDescription = "Diet Planner") },
-                    label = { Text("Diet Planner") }
-                )
-                BottomNavigationItem(
-                    selected = navController.currentDestination?.route == "settings",
-                    onClick = { navController.navigate("settings") },
-                    icon = { Icon(Icons.Default.Settings, contentDescription = "Settings") },
-                    label = { Text("Settings") }
-                )
-            }
-
-        }
-    }
-}
-
-@Composable
-fun ProfileItem(setting: String) {
-    Text(
-        text = setting,
-        style = MaterialTheme.typography.bodyMedium,
-        modifier = Modifier
-            .padding(vertical = 8.dp)
-            .fillMaxWidth()
-    )
-}
-
-
-class ProfileViewModel : ViewModel() {
-    val profileItem = listOf(
-        "User Information Display",
-        "Edit Profile",
-        "Dietary Preferences",
-        "Fitness Goals",
-        "Meal Plans",
-        "Social Features"
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun ProfilePreview() {
-    val navController = rememberNavController()
-    ProfileScreen(navController, viewModel = ProfileViewModel())
 }
 
 // TopBar Component
