+++ app/src/main/java/com/example/assignment_1/ProfileScreen.kt
@@ -1,0 +1,71 @@
+package com.example.assignment_1
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material3.Button
+import androidx.compose.material3.Divider
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import androidx.lifecycle.ViewModel
+import androidx.navigation.NavController
+
+@Composable
+fun ProfileScreen(navController: NavController, viewModel: ProfileViewModel) {
+    Surface(color = MaterialTheme.colorScheme.surface) {
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(16.dp)
+        ) {
+            Text(text = "Profile", style = MaterialTheme.typography.bodyMedium)
+            Spacer(modifier = Modifier.height(16.dp))
+            LazyColumn {
+                items(viewModel.profileItem) { setting ->
+                    ProfileItem(setting = setting)
+                    Divider()
+                }
+            }
+            Spacer(modifier = Modifier.weight(1f))
+            Button(
+                onClick = { /* Handle save button click */ },
+                modifier = Modifier.align(Alignment.End)
+            ) {
+                Text(text = "Save")
+            }
+        }
+    }
+    BottomNavigationComponent(navController = navController)
+}
+
+@Composable
+fun ProfileItem(setting: String) {
+    Text(
+        text = setting,
+        style = MaterialTheme.typography.bodyMedium,
+        modifier = Modifier
+            .padding(vertical = 8.dp)
+            .fillMaxWidth()
+    )
+}
+
+class ProfileViewModel : ViewModel() {
+    val profileItem = listOf(
+        "User Information Display",
+        "Edit Profile",
+        "Dietary Preferences",
+        "Fitness Goals",
+        "Meal Plans",
+        "Social Features"
+    )
+}
