+++ app/src/main/java/com/example/assignment_1/SettingsPage.kt
@@ -1,0 +1,99 @@
+package com.example.assignment_1
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material.BottomNavigation
+import androidx.compose.material.BottomNavigationItem
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.LocalPizza
+import androidx.compose.material.icons.filled.Search
+import androidx.compose.material.icons.filled.Settings
+import androidx.compose.material3.Button
+import androidx.compose.material3.Divider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.lifecycle.ViewModel
+import androidx.navigation.NavController
+import androidx.navigation.compose.rememberNavController
+
+@Composable
+fun SettingsScreen(navController: NavController, viewModel: SettingsViewModel) {
+    Surface(color = MaterialTheme.colorScheme.surface) {
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(16.dp)
+        ) {
+            Text(text = "Settings", style = MaterialTheme.typography.titleLarge)
+            Spacer(modifier = Modifier.height(16.dp))
+            LazyColumn {
+                items(viewModel.settingsItems) { setting ->
+                    Box(
+                        modifier = Modifier
+                            .fillMaxWidth() // Fill the width of the column
+                            .height(72.dp) // Adjust the height as needed
+                    ) {
+                        SettingsItem(setting = setting)
+                    }
+                    Divider()
+                }
+            }
+            Spacer(modifier = Modifier.weight(1f))
+            Button(
+                onClick = { /* Handle save button click */ },
+                modifier = Modifier.align(Alignment.End)
+            ) {
+                Text(text = "Save")
+            }
+        }
+    }
+    BottomNavigationComponent(navController = navController)
+}
+
+@Composable
+fun SettingsItem(setting: String) {
+    Text(
+        text = setting,
+        style = MaterialTheme.typography.bodyMedium.copy(fontSize = 16.sp),
+        modifier = Modifier
+            .padding(vertical = 16.dp, horizontal = 16.dp)
+            .fillMaxWidth()
+    )
+}
+
+class SettingsViewModel : ViewModel() {
+    val settingsItems = listOf(
+        "Notification",
+        "Sound",
+        "Dark Mode",
+        "Language",
+        "Theme Color",
+        "Font Size",
+        "Privacy",
+        "Security",
+        "Data Usage",
+        "Account",
+        // Add more settings items as needed
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+fun SettingsScreenPreview() {
+    SettingsScreen(navController = rememberNavController(), viewModel = SettingsViewModel())
+}
